+++ src/game/index.ts
@@ -1,4 +1,4 @@
 export * from "./bridge";
 export * from "./createGame";
 export * from "./demoBridge";
-
+export * from "./worldNavigation";